2023/12/22.a

+++ Table.h
@@ -3,7 +3,7 @@
 #define d_y 20
 
 std::string text_table[5] {"T-Spin","Single","Double","Triple","Tetris"};
-void goto_xy(int x, int y) {COORD pos = {x, y}; SetConsoleCursorPosition(hConsole, pos);}
+void goto_xy(short x, short y) {COORD pos = {x, y}; SetConsoleCursorPosition(hConsole, pos);}
 void set_color(const unsigned short textColor) {SetConsoleTextAttribute(hConsole, textColor);}
 void qClear(std::queue<Block*>& q) {std::queue<Block*> empty; std::swap(empty, q);}
